+++ .fvm/flutter_sdk
@@ -1,0 +1,1 @@
+/Users/admin/fvm/versions/3.7.9
