+++ .fvm/fvm_config.json
@@ -1,0 +1,4 @@
+{
+  "flutterSdkVersion": "3.7.9",
+  "flavors": {}
+}
