+++ lib/screens/auth/auth_page.dart
@@ -1,7 +1,8 @@
-import 'package:blog_app/blog_post/page1.dart';
-import 'package:blog_app/sign_up.dart';
+import 'package:blog_app/screens/auth/sign_up.dart';
 import 'package:firebase_auth/firebase_auth.dart';
 import 'package:flutter/material.dart';
+
+import '../blog_post/page1.dart';
 
 class AuthPage extends StatefulWidget {
   const AuthPage({super.key});
@@ -18,8 +19,7 @@ class _AuthPageState extends State<AuthPage> {
           stream: FirebaseAuth.instance.authStateChanges(),
           builder: (context, snapshot) {
             if (snapshot.hasData) {
-              return const PageOne(); }
-              
+              return const PageOne(); }   
               else {
                 return const SignupPage();
               }
