+++ lib/screens/auth/login.dart
@@ -1,0 +1,167 @@
+
+import 'package:blog_app/screens/auth/sign_up.dart';
+import 'package:flutter/material.dart';
+
+import '../../services/auth_services.dart';
+
+class LoginScreen extends StatefulWidget {
+  const LoginScreen({super.key});
+
+  @override
+  State<LoginScreen> createState() => _LoginScreenState();
+}
+
+final formKey = GlobalKey<FormState>();
+final emailController = TextEditingController();
+final passwordController = TextEditingController();
+
+@override
+void dispose() {
+  emailController.dispose();
+  passwordController.dispose();
+}
+
+class _LoginScreenState extends State<LoginScreen> {
+  @override
+  Widget build(BuildContext context) {
+    return Stack(
+      children: [
+        ShaderMask(
+          shaderCallback: (rect) => const LinearGradient(
+            begin: Alignment.bottomCenter,
+            end: Alignment.center,
+            colors: [Colors.black, Colors.transparent],
+          ).createShader(rect),
+          blendMode: BlendMode.darken,
+          child: Container(
+            decoration: const BoxDecoration(
+              image: DecorationImage(
+                image: AssetImage("images/pexels-cottonbro-studio-5076522.jpg"),
+                fit: BoxFit.cover,
+                colorFilter: ColorFilter.mode(Colors.black54, BlendMode.darken),
+              ),
+            ),
+          ),
+        ),
+        Scaffold(
+          backgroundColor: Colors.transparent,
+          body: Form(
+            key: formKey,
+            child: Column(
+              children: [
+                const Flexible(
+                  child: Center(
+                    child: Text(
+                      "Everyday blog",
+                      style: TextStyle(
+                          color: Colors.white,
+                          fontSize: 50,
+                          fontWeight: FontWeight.bold),
+                    ),
+                  ),
+                ),
+                const TextInputField(),
+                Padding(
+                  padding: const EdgeInsets.symmetric(vertical: 10.0),
+                  child: Container(
+                    height: 60,
+                    width: 320,
+                    decoration: BoxDecoration(
+                      color: Colors.grey[500]?.withOpacity(0.5),
+                      borderRadius: BorderRadius.circular(16),
+                    ),
+                    child: TextFormField(
+                      controller: passwordController,
+                      decoration: const InputDecoration(
+                        border: InputBorder.none,
+                        prefixIcon: Padding(
+                          padding: EdgeInsets.symmetric(horizontal: 20.0),
+                          child: Icon(
+                            Icons.lock,
+                            size: 28,
+                            color: Colors.white,
+                          ),
+                        ),
+                        hintText: "Password",
+                        hintStyle: TextStyle(fontSize: 22, color: Colors.white),
+                      ),
+                      obscureText: true,
+                      style: const TextStyle(fontSize: 22, color: Colors.white),
+                      keyboardType: TextInputType.text,
+                      textInputAction: TextInputAction.done,
+                    ),
+                  ),
+                ),
+                Align(
+                  alignment: Alignment.centerRight,
+                  child: TextButton(
+                    onPressed: () {},
+                    child: const Text(
+                      "Forgot Password?",
+                      style: TextStyle(
+                          fontSize: 19,
+                          color: Colors.white,
+                          fontWeight: FontWeight.w300),
+                    ),
+                  ),
+                ),
+                const SizedBox(
+                  height: 15,
+                ),
+                Container(
+                  height: 60,
+                  width: 320,
+                  decoration: BoxDecoration(
+                      borderRadius: BorderRadius.circular(16),
+                      color: Colors.blue),
+                  child: TextButton(
+                    onPressed: (){
+                      final isValid = formKey.currentState!.validate();
+                      if (!isValid) return;
+                      AuthServices.login(
+                        emailController: emailController,
+                        passwordController: passwordController
+                      );
+                      },
+                    child: Text(
+                      "Login",
+                      style: TextStyle(
+                          fontSize: 22,
+                          color: Colors.white,
+                          fontWeight: FontWeight.w900),
+                    ),
+                  ),
+                ),
+                const SizedBox(
+                  height: 15,
+                ),
+                Align(
+                  alignment: Alignment.center,
+                  child: TextButton(
+                    onPressed: () {
+                      Navigator.push(
+                          context,
+                          MaterialPageRoute(
+                            builder: (context) => const SignupPage(),
+                          ));
+                    },
+                    child: const Text(
+                      "Create New Account",
+                      style: TextStyle(
+                          fontSize: 20,
+                          color: Colors.white,
+                          fontWeight: FontWeight.w300),
+                    ),
+                  ),
+                ),
+                const SizedBox(
+                  height: 15,
+                ),
+              ],
+            ),
+          ),
+        )
+      ],
+    );
+  }
+}
